rename slot files
