+++ src/components/items/card-redesign/itemCard.Footer.tsx
