+++ src/components/items/card-redesign/itemCard.Source.test.tsx
@@ -3,17 +3,17 @@ import { describe, expect, it } from "vitest"
 
 import { ThemeWrapper } from "#testUtils/renderUtils.tsx"
 
-import { ItemCardSource } from "./itemCardSource.tsx"
+import { ItemCardSource } from "./itemCard.Source.tsx"
 
 describe("ItemCardSource", () => {
   it("renders its children", () => {
-    render(<ItemCardSource>Street Grimoire p.42</ItemCardSource>, { wrapper: ThemeWrapper })
+    render(<ItemCardSource source={{ book: "SR4A", page: 427 }} />, { wrapper: ThemeWrapper })
 
-    expect(screen.getByText("Street Grimoire p.42")).toBeDefined()
+    expect(screen.getByText("SR4A p.427")).toBeDefined()
   })
 
   it("renders nothing when there are no children", () => {
-    const { container } = render(<ItemCardSource />, { wrapper: ThemeWrapper })
+    const { container } = render(<ItemCardSource source={null} />, { wrapper: ThemeWrapper })
 
     expect(container.firstChild).toBeNull()
   })
