+++ src/components/items/types/vehicles/vehicleItemCard.tsx
@@ -4,7 +4,7 @@ import type { FC } from "react"
 import { GearItemCard } from "#/components/items/card/gearItemCard.tsx"
 import { ItemCard } from "#/components/items/card/itemCard.tsx"
 import { ItemStatChip } from "#/components/items/card/itemStatChip.tsx"
-import { ItemCardDamageTrack } from "#/components/items/card-redesign/itemCardDamageTrack.tsx"
+import { ItemCardDamageTrack } from "#/components/items/card-redesign/itemCard.DamageTrack.tsx"
 import { GenericItemCard } from "#/components/items/genericItemCard.tsx"
 import { Nuyen } from "#/components/ui/nuyen.tsx"
 import type { VehicleData } from "#/system/gear/vehicleData.ts"
