+++ src/lib/slotUtils.ts
@@ -1,0 +1,8 @@
+import type { FC, ReactElement, ReactNode } from "react"
+import { isValidElement } from "react"
+
+export function isElementType<TProps>(elementType: FC<TProps>) {
+  return (item: ReactNode): item is ReactElement<TProps> => {
+    return isValidElement(item) && item.type === elementType
+  }
+}
